Various improvements and end-to-end tests

+++ exporters/fluentd/src/recordable.cc
@@ -1,18 +1,5 @@
-/*
- * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+/// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
 
 #include "opentelemetry/exporters/fluentd/recordable.h"
 
@@ -234,7 +221,7 @@ void Recordable::SetStartTime(opentelemetry::common::SystemTimestamp start_time)
 {
   options_[FLUENT_FIELD_STARTTIME] =
       get_msgpack_eventtimeext(
-        std::chrono::duration_cast<std::chrono::seconds>(start_time.time_since_epoch()).count(),
+        static_cast<int32_t>(std::chrono::duration_cast<std::chrono::seconds>(start_time.time_since_epoch()).count()),
         std::chrono::duration_cast<std::chrono::nanoseconds>(start_time.time_since_epoch()).count() % 1000000000);
 }
